Adds strictProperties and $merge
For draft 5 of jsonschema, there is a proposal to create "$merge" and
"strictProperties". A description of these can be found at:

    https://github.com/fge/json-schema-validator/wiki/v5:-merge

        and

    https://github.com/fge/json-schema-validator/wiki/v5:-strictProperties

This implementation differs from the description in one way: the
description specifies "merge" whereas I have used "$merge" following the
more recent discussions on the mailing list:

    https://groups.google.com/forum/#!forum/jsonschema

This is NOT a compliant draft 5 implementation; draft 5 of jsonschema
has not been published yet. At best this is a subset of what may become
draft 5 of jsonschema.